# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Campus Online API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
  studentId?: string;
  department?: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  student: "student",
  club_official: "club_official",
  admin: "admin",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  studentId?: string | null;
  department?: string | null;
  avatarUrl?: string | null;
  isActive: boolean;
  clubId?: number | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface UpdateUserBody {
  name?: string;
  department?: string;
  avatarUrl?: string;
  isActive?: boolean;
}

export interface UserListResponse {
  users: User[];
  total: number;
  page: number;
  limit: number;
}

export interface Club {
  id: number;
  name: string;
  description?: string | null;
  category?: string | null;
  logoUrl?: string | null;
  memberCount: number;
  eventCount: number;
  createdAt: string;
}

export interface CreateClubBody {
  name: string;
  description?: string;
  category?: string;
  logoUrl?: string;
}

export interface UpdateClubBody {
  name?: string;
  description?: string;
  category?: string;
  logoUrl?: string;
}

export interface ClubListResponse {
  clubs: Club[];
  total: number;
  page: number;
  limit: number;
}

export type EventStatus = (typeof EventStatus)[keyof typeof EventStatus];

export const EventStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export type EventUserReaction =
  | (typeof EventUserReaction)[keyof typeof EventUserReaction]
  | null;

export const EventUserReaction = {
  going: "going",
  interested: "interested",
} as const;

export interface Event {
  id: number;
  title: string;
  description?: string | null;
  category?: string | null;
  location?: string | null;
  imageUrl?: string | null;
  startDate: string;
  endDate?: string | null;
  status: EventStatus;
  clubId: number;
  clubName?: string | null;
  attendeeCount: number;
  interestedCount: number;
  userReaction?: EventUserReaction;
  createdAt: string;
}

export interface CreateEventBody {
  title: string;
  description?: string;
  category?: string;
  location?: string;
  imageUrl?: string;
  startDate: string;
  endDate?: string;
}

export type UpdateEventBodyStatus =
  (typeof UpdateEventBodyStatus)[keyof typeof UpdateEventBodyStatus];

export const UpdateEventBodyStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface UpdateEventBody {
  title?: string;
  description?: string;
  category?: string;
  location?: string;
  imageUrl?: string;
  startDate?: string;
  endDate?: string;
  status?: UpdateEventBodyStatus;
}

export interface EventListResponse {
  events: Event[];
  total: number;
  page: number;
  limit: number;
}

export type EventReactionType =
  (typeof EventReactionType)[keyof typeof EventReactionType];

export const EventReactionType = {
  going: "going",
  interested: "interested",
} as const;

export interface EventReaction {
  id: number;
  userId: number;
  eventId: number;
  type: EventReactionType;
  userName?: string | null;
  createdAt: string;
}

export type ReactToEventBodyType =
  (typeof ReactToEventBodyType)[keyof typeof ReactToEventBodyType];

export const ReactToEventBodyType = {
  going: "going",
  interested: "interested",
} as const;

export interface ReactToEventBody {
  type: ReactToEventBodyType;
}

export interface Announcement {
  id: number;
  title: string;
  content: string;
  clubId?: number | null;
  clubName?: string | null;
  isGlobal: boolean;
  authorId: number;
  authorName?: string | null;
  createdAt: string;
}

export interface CreateAnnouncementBody {
  title: string;
  content: string;
  isGlobal?: boolean;
}

export interface AnnouncementListResponse {
  announcements: Announcement[];
  total: number;
  page: number;
  limit: number;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface DashboardSummary {
  totalEvents: number;
  upcomingEvents: number;
  totalClubs: number;
  totalStudents: number;
  pendingApprovals: number;
  recentAnnouncements: number;
  topCategories: CategoryCount[];
}

export type ListUsersParams = {
  page?: number;
  limit?: number;
  search?: string;
  role?: ListUsersRole;
  isActive?: boolean;
};

export type ListUsersRole = (typeof ListUsersRole)[keyof typeof ListUsersRole];

export const ListUsersRole = {
  student: "student",
  club_official: "club_official",
  admin: "admin",
} as const;

export type ListClubsParams = {
  page?: number;
  limit?: number;
  search?: string;
};

export type ListEventsParams = {
  page?: number;
  limit?: number;
  search?: string;
  category?: string;
  clubId?: number;
  status?: ListEventsStatus;
  upcoming?: boolean;
};

export type ListEventsStatus =
  (typeof ListEventsStatus)[keyof typeof ListEventsStatus];

export const ListEventsStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export type ListAnnouncementsParams = {
  page?: number;
  limit?: number;
  clubId?: number;
};