# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Campus Online API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /clubs:
    get:
      operationId: listClubs
      tags: [clubs]
      summary: List all clubs
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: List of clubs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClubListResponse"
    post:
      operationId: createClub
      tags: [clubs]
      summary: Create a club (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClubBody"
      responses:
        "201":
          description: Created club
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Club"
...
  /clubs/{id}:
    get:
      operationId: getClub
      tags: [clubs]
      summary: Get club by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Club details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Club"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateClub
      tags: [clubs]
      summary: Update club
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClubBody"
      responses:
        "200":
          description: Updated club
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Club"
...
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List announcements
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: clubId
          schema:
            type: integer
      responses:
        "200":
          description: List of announcements
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnouncementListResponse"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create announcement (club official or admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnnouncementBody"
      responses:
        "201":
          description: Created announcement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{id}:
    get:
      operationId: getAnnouncement
      tags: [announcements]
      summary: Get announcement by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Announcement details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      summary: Delete announcement (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        studentId:
          type: string
        department:
          type: string
      required:
        - name
        - email
        - password
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, club_official, admin]
        studentId:
          type: string
          nullable: true
        department:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        clubId:
          type: integer
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - isActive
        - createdAt
...
    Club:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
        memberCount:
          type: integer
        eventCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - memberCount
        - eventCount
        - createdAt
    CreateClubBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        logoUrl:
          type: string
      required:
        - name
    UpdateClubBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        logoUrl:
          type: string
    ClubListResponse:
      type: object
      properties:
        clubs:
          type: array
          items:
            $ref: "#/components/schemas/Club"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - clubs
        - total
        - page
        - limit
...
    Announcement:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        clubId:
          type: integer
          nullable: true
        clubName:
          type: string
          nullable: true
        isGlobal:
          type: boolean
        authorId:
          type: integer
        authorName:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - title
        - content
        - isGlobal
        - authorId
        - createdAt
    CreateAnnouncementBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        isGlobal:
          type: boolean
      required:
        - title
        - content
    AnnouncementListResponse:
      type: object
      properties:
        announcements:
          type: array
          items:
            $ref: "#/components/schemas/Announcement"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - announcements
        - total
        - page
        - limit
...