# lib/api-zod/src/generated/types/club.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Campus Online API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Club {
  id: number;
  name: string;
  description?: string | null;
  category?: string | null;
  logoUrl?: string | null;
  memberCount: number;
  eventCount: number;
  createdAt: string;
}