# lib/api-zod/src/generated/types/updateClubBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Campus Online API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateClubBody {
  name?: string;
  description?: string;
  category?: string;
  logoUrl?: string;
}