# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Campus Online API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  studentId?: string | null;
  department?: string | null;
  avatarUrl?: string | null;
  isActive: boolean;
  clubId?: number | null;
  createdAt: string;
}